Promise variable in init shouldn't be on Plugins
Currently, the Promise in the start function in the Cordova definition is created off the plugins settings and then chained off of by several other things. However, an app may not use plugins, in which case errors will occur. Platforms are required, which makes them a more suitable candidate to start the promise chain. I've prototyped that here, just to give an idea of how this might work, though it is likely not exactly right (I'm doing it in the browser, so I can't see if it's working).